Make sure that DevToolsHttpHandlerImpl is always destroyed on the UI thread.
If we post back to the UI thread from StopHandlerThread, it can happen that the task is executed before we're finished on the FILE thread, which means we drop the last reference on the FILE thread.

Using PostTaskAndReply avoids that, because it makes sure that the task is always destroyed on the originating thread.

BUG=161027


Review URL: https://chromiumcodereview.appspot.com/11434003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@170291 0039d316-1c4b-4281-b951-d872f2087c98