Move initialization of D3D earlier in present thread lifetime.
This improves startup time.

Fixed bug where initial window size was treated as a resize and triggered the slow GDI path.

Previous version was r171174, which was reverted. This patch is different because it does not initialize D3D until an AcceleratedPresenter is created but does not wait for the first present.

BUG=163215,164414,164398

Review URL: https://chromiumcodereview.appspot.com/11445029

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@171413 0039d316-1c4b-4281-b951-d872f2087c98