Attempt to linearize the scroll deltas created in RenderWidgetHostViewBase
We're currently shifting 10 pixels every time the smooth scroll gesture gets
serviced. If we get serviced at uneven intervals (which seems to happen), then
we get uneven scroll speeds. This patch attempts to force the scroll position
to be a linear function of time, though there may be jank if we generate the
events an uneven intervals.

BUG=162779


Review URL: https://chromiumcodereview.appspot.com/11416283

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@171506 0039d316-1c4b-4281-b951-d872f2087c98