[Android WebView] Add new chrome/-less pak dependency.
android_webview must not depend on the chrome/ layer, so
we want to remove the last dependencies we have for generating resource paks. Toward this, define a new android_webview pak that depends only on content/ resources. 

A later change will remove the dependency on the paks that contain chrome/resources.
    
For now we use the same resources Content Shell does. It may be possible to trim this further, but seems like a good starting point.

Android only, Android bots green
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/11348326

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@171522 0039d316-1c4b-4281-b951-d872f2087c98