immersive-mode: Fix hover-highlight for frame buttons.
When the tabstrip etc. top-views are 'revealed' in immersive mode, the
RevealView takes care of painting the frame buttons. But paint-updates in
the frame buttons don't take effect because the paint-requests from these
buttons do not reach the RevealView. So, when in revealing state, redirect
the paint-requests to the RevealView appropriately.

BUG=163637

Review URL: https://codereview.chromium.org/11543012

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@172628 0039d316-1c4b-4281-b951-d872f2087c98