Track managed memory usage in the command buffer.
This adds infrastructure necessary to support tracking which allocations
are through the compositor (or, more generally, any managed allocator that
the GPU memory manager affect) and allocations that are not (e.g, WebGL).

This is part of a scheme which will add a GL extension where the compositor
can mark a texture as being managed by a call to glTexParameteri. 

This information had previously been tracked by an IPC being sent by the
compositor to the GPU process informing it of the managed memory usage (in
addition to the usage required for all visible content, etc). This had
the two problems. First, the information was not accurate -- the data sent
about managed allocations to the GPU memory manager would always lag the
actual allocations. Second, these IPCs were expensive and needed to be
throttled. The more we throttle the IPCs, the less accurate the information
would become, making the GPU memory manager's behavior less precise and
predictable.

BUG=164947


Review URL: https://chromiumcodereview.appspot.com/11516014

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@172754 0039d316-1c4b-4281-b951-d872f2087c98