New icons for Settings App in app launcher (and Windows taskbar).
This introduces a new designer icon for the Settings App with a
transparent margin so that it doesn't appear oversized. New assets
are also introduced -- high-quality pre-resized at 48x48px for the
icon in the app launcher, and 32x32 for the Windows 7 "large icons"
taskbar.

BUG=130210
TEST=Settings app icon updated to be slightly smaller, and now with
     8 cogs instead of 6.


Review URL: https://chromiumcodereview.appspot.com/11488025

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@172846 0039d316-1c4b-4281-b951-d872f2087c98