Read the ChromeOS channel info earlier during startup.
This makes the channel available much earlier, which 
allows users of that info, such as the VariationsService 
to work correctly. 

This CL also refactors StatisticsProvider initialization 
to be started explicitly, instead of as a result of the 
first call to GetInstance(). 

Previously, the initialization ended up happening as a 
result of BrowserPolicyConnector::CompleteInitialization() 
which was triggered by BrowserPolicyConnector::Init() when 
the local state was being initialized. 

BUG=107333
TEST=Manual, by instrumenting code and once this lands, checking that channel-targeted Finch studies get applied on ChromeOS.

Review URL: https://chromiumcodereview.appspot.com/11549025

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@173507 0039d316-1c4b-4281-b951-d872f2087c98