Forward Webkit Gesture events to interested RenderViewObservers.
The ChromeRenderViewObserver overrides the RenderViewObserver::DidHandleGestureEvent virtual and
forwards the ChromeViewHostMsg_FocusedEditableNodeTouched IPC notification to the brower if the user
tapped on an editable field on the page. 

We will be handling this notification in the browser to display the Windows 8 on screen keyboard. That
in a subsequent CL.

BUG=166516
R=jam
Review URL: https://codereview.chromium.org/11622009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@173601 0039d316-1c4b-4281-b951-d872f2087c98