Add iOS-only stub implementation of static function test_launcher::GetCurrentTestLauncherDelegate.
The stub static function returns NULL, as there is no such delegate on iOS.  Because it's a static function, there is no need to worry about drift within the class.  The file where the function is called already checks against a NULL delegate, and this avoids an awkward #if-def block.

BUG=None


Review URL: https://chromiumcodereview.appspot.com/10908118

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@173754 0039d316-1c4b-4281-b951-d872f2087c98