Fix a problem that we are showing webpage as background before video starts playing
getDefaultSize() will return the screen width and screen height. So the default size of the surfaceview will occupy the full screen before we get the actual video size.
Using an empty size in this case so that we will show a black background with spinning circle instead of the webpage.

BUG=166417


Review URL: https://chromiumcodereview.appspot.com/11614019

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@173802 0039d316-1c4b-4281-b951-d872f2087c98