Clip the header of ash windows to the frame height
Anything painted below the frame are becomes visible if the app enables
the transparent background feature.

For lack of a better way to compute the actual frame height, we clip the
drawing to the height of the close button.

I think ScopedCanvas would come useful here, but it's currently private
to view.cc. If it seems a good idea, I could move it to canvas.h.

BUG=152838
TEST=visually confirmed with patch for bug 125295 applied


Review URL: https://chromiumcodereview.appspot.com/11087065

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@173823 0039d316-1c4b-4281-b951-d872f2087c98