Fix Linux capturer crash on resize
Problem was a race-condition triggered by two resize
operations in quick succession:

* Capturer receives ConfigureNotify with a large screen size.
It stores the large size, and calls XServerPixelBuffer::Init()
* Meanwhile, the root window is resized to a smaller size.
* XServerPixelBuffer::Init() directly queries the current size,
and creates buffers with the new smaller size.
* Capturer requests screen capture, using the stored large size.
It crashes when trying to copy too much data from the
PixelBuffer.
* If it didn't crash, the Capturer would then get the
ConfigureNotify event for the smaller size.

This CL fixes the XServerPixelBuffer to use the sizes from the
ConfigureNotify events. As an added precaution, this also clips
the XDamage regions to the current size of the PixelBuffer,
before capturing screen pixels.


BUG=166713
TEST="while true; do xrandr -s 1; xrandr -s 2; done"


Review URL: https://chromiumcodereview.appspot.com/11639025

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@174112 0039d316-1c4b-4281-b951-d872f2087c98