Remove DCHECK for PSS::SetEncryptionPassphrase
For Android, we do not sync passwords, so the cryptographer is not
initialized on a non-encrypted account.

The |passphrase_required_reason_| is therefore set to REASON_ENCRYPTION
at the time of the DCHECK if you turn on custom passphrase encryption
after already being signed in and syncing data.

BUG=165309


Review URL: https://chromiumcodereview.appspot.com/11637045

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@174351 0039d316-1c4b-4281-b951-d872f2087c98