sync: Move unique_client_tag hashing code
This is one of many changes that will lead up to the introduction of
the bookmark UniquePosition algorithm.

We eventually want to use the hashing code to create a unique tag for
bookmarks.  This tag will be initialized at bookmark creation time in
MutableEntry.  This patch moves the hashing code out of BaseNode and
into syncable_util so it can be accessed from the syncable layer.

BUG=145412,126505


Review URL: https://chromiumcodereview.appspot.com/11571092

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@174449 0039d316-1c4b-4281-b951-d872f2087c98