Fix Debian build script for svn-based checkout
In an svn-based checkout, the git commands for
getting the SVN revision failed and caused the
script to exit silently.

With this change, the script will do the build
without adding the revision number info, which
isn't needed anyway for an official build since
the version number uniquely identifies it.

NOTRY=true


Review URL: https://chromiumcodereview.appspot.com/11624032

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@174634 0039d316-1c4b-4281-b951-d872f2087c98