Prevent interleaving of error messages with perf result output.
On the perf bots we are seeing a LOG(ERROR) line being interleaved with the
perf result output. This causes the perf result to be improperly parsed.

I'm hoping that this interleaving is due to stdout buffering and these flushes
will reduce the chance that any other output will be interleaved. But I haven't
directly verified the fix as the problem is flaky and difficult to reproduce.

BUG=166576,155897
TEST=None


Review URL: https://chromiumcodereview.appspot.com/11637050

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@175059 0039d316-1c4b-4281-b951-d872f2087c98