mac: Move tabstrip highlight drawing from code to bitmap
Inactive tabs used to draw a white and a grey highlight line at the bottom.
https://codereview.chromium.org/11280168 moved the border drawing to a bitmap,
and the bitmap didn't include the white highlight line, so that changelist also
made the line of the tabstrip (the line at the top of the toolbar, visible in
the parts of the window where no tabs sit on top of the toolbar) darker.

The new plan is to let this border always been drawn by the tab strip (via a
bitmap) and change background tabs to not draw on top of that border. This
will match the CrOS drawing approach. This change changes the border drawn by
the tab strip to
a) be bitmap-based
b) be drawn on top of the toolbar bitmap, so that the bitmap can be made
   partially transparent in the future.

The current border highlight doesn't have transparency, so this CL should
have no visible effect yet. A future CL will change the tab border bitmaps to
not include the tab strip border, and will change tab_view.mm to not draw the
tab fill of inactive bitmaps on top of the border drawn by the tabstrip.

Based on a patch from sail@chromium.org

BUG=166403
TEST=nothing changes

Review URL: https://codereview.chromium.org/11743029

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@175064 0039d316-1c4b-4281-b951-d872f2087c98