[Sync] Add support for loading, updating and querying delete journals in Directory.
Delete journals keeps deleted metas until the persistence of the deletes in
native model is confirmed. When an entry is deleted in sync model, a copy is 
added to delete journals in memory and saved in database later. Next time when
the client restarts, if some native data doesn't match with sync data but
matches with a delete journal, it's safe to assume that it's because native
delete was not persisted and the native data should be deleted. This helps
prevent back-from-dead problem due to native model and sync model get
out-of-sync.

BUG=121928

Review URL: https://chromiumcodereview.appspot.com/11441026

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@175248 0039d316-1c4b-4281-b951-d872f2087c98