Reverting fix for crbug.com/138309 which caused this problem (Filed bug crbug.com/168910 to re-fix 138309 differently).
The fix for crbug.com/138309 dismisses context menus on ChromeOS when Ctrl or Alt is pressed.
This prevents context menus that allow for Ctrl clicks to work (because they are closed prematurely).

The problem that the fix for 138309 tried to address is that "tab navigation" short-cuts work but leave
context menus up. These types of shortcuts should either be blocked or dismiss the context menu.

For more info see:
crbug.com/167539 (this bug)
crbug.com/139308 (bug fix being reverted)
crbug.com/168910 (replica of 139308)

BUG=167539
TEST=manual:
Create bookmark folder with entries
Click on bookmark folder 
Control Click on an entry

Review URL: https://chromiumcodereview.appspot.com/11826012

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@175715 0039d316-1c4b-4281-b951-d872f2087c98